requirements.txt
bleak
pyyaml
ultralytics
opencv-python
numpy
matplotlib
python-dotenv
